Simplified event override pruning - moved onto Event struct
