+++ redical_core/src/lib.rs
@@ -7,7 +7,6 @@ mod event_occurrence_override;
 mod geo_index;
 mod inverted_index;
 mod utils;
-mod prune;
 
 #[cfg(test)]
 mod testing;
@@ -21,6 +20,5 @@ pub use event_occurrence_override::*;
 pub use geo_index::*;
 pub use inverted_index::*;
 pub use utils::*;
-pub use prune::*;
 
 pub mod queries;
