+++ redical_redis/src/commands/rdcl_evo_prune.rs
@@ -1,9 +1,8 @@
 use std::str::FromStr;
-use std::ops::Bound;
 
 use redis_module::{Context, NextArg, NotifyEvent, RedisError, RedisResult, RedisString, Status, RedisValue};
 
-use redical_core::{Calendar, EventOccurrenceOverride, CalendarIndexUpdater, InvertedEventIndex, prune_event_overrides};
+use redical_core::{Calendar, EventOccurrenceOverride, CalendarIndexUpdater, InvertedEventIndex};
 use crate::datatype::CALENDAR_DATA_TYPE;
 
 use redical_ical::values::date_time::DateTime;
@@ -22,7 +21,7 @@ fn notify_keyspace_event(ctx: &Context, calendar_uid: &RedisString, event_uid: &
     Ok(())
 }
 
-fn prune_calendar_events_overrides(calendar: &mut Calendar, event_uid: String, from_timestamp: i64, until_timestamp: i64, callback: impl FnMut(i64, EventOccurrenceOverride)) -> Result<(), RedisError> {
+fn prune_calendar_events_overrides(calendar: &mut Calendar, event_uid: String, from_timestamp: i64, until_timestamp: i64) -> Result<Vec<(i64, EventOccurrenceOverride)>, RedisError> {
     let Some(mut event) = calendar.events.get(&event_uid).cloned() else {
         return Err(RedisError::String(format!(
             "No event with UID: '{}' found",
@@ -30,12 +29,7 @@ fn prune_calendar_events_overrides(calendar: &mut Calendar, event_uid: String, f
         )));
     };
 
-    prune_event_overrides(
-        event.as_mut(),
-        Bound::Included(from_timestamp),
-        Bound::Included(until_timestamp),
-        callback,
-    ).map_err(RedisError::String)?;
+    let removed_overrides = event.as_mut().prune_event_overrides(from_timestamp, until_timestamp).map_err(RedisError::String)?;
 
     let event_uid = event.uid.uid.to_string();
 
@@ -96,7 +90,7 @@ fn prune_calendar_events_overrides(calendar: &mut Calendar, event_uid: String, f
             .map_err(|error| RedisError::String(error.to_string()))?;
     }
 
-    Ok(())
+    Ok(removed_overrides)
 }
 
 fn timestamp_from_date_string(date_string: &str) -> Result<i64, RedisError> {
@@ -135,8 +129,6 @@ pub fn redical_event_override_prune(ctx: &Context, args: Vec<RedisString>) -> Re
         )));
     };
 
-    let mut removed_event_occurrence_overrides: Vec<(String, String)> = Vec::new();
-
     if args.len() == 3 {
         let event_uid = args.next_arg()?.to_string();
 
@@ -150,20 +142,14 @@ pub fn redical_event_override_prune(ctx: &Context, args: Vec<RedisString>) -> Re
             format!("rdcl.evo_prune: calendar_uid: {calendar_uid} event_uid: {event_uid} from_timestamp: {from_timestamp} until_timestamp: {until_timestamp}").as_str()
         );
 
-        prune_calendar_events_overrides(
-            calendar,
-            event_uid.to_owned(),
-            from_timestamp,
-            until_timestamp,
-            |override_timestamp, _event_occurrence_override| {
-                removed_event_occurrence_overrides.push(
-                    (
-                        event_uid.clone(),
-                        DateTime::from(override_timestamp).render_formatted_date_time(None),
-                    )
-                );
-            }
-        )?;
+        // Use this command when replicating across other Redis instances.
+        // We call this here to ensure all replicas begin at and reach the same point if any errors
+        // are raised in the following prune process.
+        ctx.replicate_verbatim();
+
+        for (override_timestamp, _event_occurrence_override) in prune_calendar_events_overrides(calendar, event_uid.to_owned(), from_timestamp, until_timestamp)? {
+            notify_keyspace_event(ctx, &calendar_uid, &event_uid, &DateTime::from(override_timestamp).render_formatted_date_time(None))?;
+        }
     } else {
         let (from_timestamp, until_timestamp) =
             timestamps_from_date_strings(
@@ -178,29 +164,16 @@ pub fn redical_event_override_prune(ctx: &Context, args: Vec<RedisString>) -> Re
         // TODO: Inefficient - optimise towards copy-less approach.
         let event_uids: Vec<String> = calendar.events.keys().map(String::from).collect();
 
+        // Use this command when replicating across other Redis instances.
+        // We call this here to ensure all replicas begin at and reach the same point if any errors
+        // are raised in the following prune process.
+        ctx.replicate_verbatim();
+
         for event_uid in event_uids {
-            prune_calendar_events_overrides(
-                calendar,
-                event_uid.to_owned(),
-                from_timestamp,
-                until_timestamp,
-                |override_timestamp, _event_occurrence_override| {
-                    removed_event_occurrence_overrides.push(
-                        (
-                            event_uid.clone(),
-                            DateTime::from(override_timestamp).render_formatted_date_time(None),
-                        )
-                    );
-                }
-            )?;
+            for (override_timestamp, _event_occurrence_override) in prune_calendar_events_overrides(calendar, event_uid.to_owned(), from_timestamp, until_timestamp)? {
+                notify_keyspace_event(ctx, &calendar_uid, &event_uid, &DateTime::from(override_timestamp).render_formatted_date_time(None))?;
+            }
         }
-    }
-
-    // Use this command when replicating across other Redis instances.
-    ctx.replicate_verbatim();
-
-    for (event_uid, override_date_string) in removed_event_occurrence_overrides {
-        notify_keyspace_event(ctx, &calendar_uid, &event_uid, &override_date_string)?;
     }
 
     Ok(RedisValue::Bool(true))
